Use new alert view controller in SendHodler module
- remove old AlertViewController
