--- BankWallet/BankWallet/Modules/Send/SubModules/SendHodler/LockTimeInterval/LockTimeIntervalModule.swift
@@ -1,9 +1,0 @@
-import Hodler
-
-protocol ILockTimeIntervalRouter {
-    func dismiss(with: HodlerPlugin.LockTimeInterval?)
-}
-
-protocol ILockTimeIntervalDelegate: class {
-    func onSelect(lockTimeInterval: HodlerPlugin.LockTimeInterval?)
-}
