+++ BankWallet/BankWallet/Modules/Send/SubModules/SendHodler/SendHodlerPresenter.swift
@@ -29,17 +29,11 @@ extension SendHodlerPresenter: ISendHodlerModule {
 extension SendHodlerPresenter: ISendHodlerViewDelegate {
 
     func onLockTimeIntervalSelectorTap() {
-        router.openLockTimeIntervals(selected: lockTimeInterval, lockTimeIntervalDelegate: self)
-    }
-
-}
-
-extension SendHodlerPresenter: ILockTimeIntervalDelegate {
-
-    func onSelect(lockTimeInterval: HodlerPlugin.LockTimeInterval?) {
-        self.lockTimeInterval = lockTimeInterval
-        view?.setLockTimeInterval(lockTimeInterval: lockTimeInterval)
-        delegate?.onUpdateLockTimeInterval()
+        router.openLockTimeIntervals(selected: lockTimeInterval) { [weak self] selectedInterval in
+            self?.lockTimeInterval = selectedInterval
+            self?.view?.setLockTimeInterval(lockTimeInterval: selectedInterval)
+            self?.delegate?.onUpdateLockTimeInterval()
+        }
     }
 
 }
