+++ BankWallet/BankWallet/Modules/Send/SubModules/SendHodler/SendHodlerRouter.swift
@@ -21,10 +21,24 @@ extension SendHodlerRouter {
 
 extension SendHodlerRouter: ISendHodlerRouter {
 
-    func openLockTimeIntervals(selected: HodlerPlugin.LockTimeInterval?, lockTimeIntervalDelegate: ILockTimeIntervalDelegate) {
-        LockTimeIntervalRouter.module(lockTimeIntervalDelegate: lockTimeIntervalDelegate, lockTimeInterval: selected).map { viewController in
-            self.viewController?.present(viewController, animated: true)
+    func openLockTimeIntervals(selected: HodlerPlugin.LockTimeInterval?, onSelect: @escaping (HodlerPlugin.LockTimeInterval?) -> ()) {
+        var intervals: [HodlerPlugin.LockTimeInterval?] = [nil]
+
+        HodlerPlugin.LockTimeInterval.allCases.forEach { interval in
+            intervals.append(interval)
         }
+
+        let alertController = AlertViewController(
+                header: "send.hodler_locktime".localized,
+                rows: intervals.map { interval in
+                    let title = interval.map { $0.title } ?? "send.hodler_locktime_off".localized
+                    return AlertRow(text: title, selected: interval == selected)
+                }
+        ) { selectedIndex in
+            onSelect(intervals[selectedIndex])
+        }
+
+        viewController?.present(alertController, animated: true)
     }
 
 }
