+++ BankWallet/BankWallet/en.lproj/Localizable.strings
@@ -138,10 +138,10 @@ Go to Settings - > Unstoppable and allow access to Camera.";
 "send.next_button" = "Next";
 "send.error.invalid_address" = "Invalid Address";
 "send.hodler_locktime" = "Lock Time";
-"send.hodler_locktime_hour" = "Hour";
-"send.hodler_locktime_month" = "Month";
+"send.hodler_locktime_hour" = "1 hour";
+"send.hodler_locktime_month" = "1 month";
 "send.hodler_locktime_half_year" = "6 months";
-"send.hodler_locktime_year" = "Year";
+"send.hodler_locktime_year" = "1 year";
 "send.hodler_locktime_off" = "Off";
 
 "send.confirmation.title" = "Confirm";
